requirements.txt
# Core API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0        # kept for tests (conftest uses sqlite+aiosqlite:///:memory:)
asyncpg>=0.29.0          # PostgreSQL + TimescaleDB async driver

# Blockchain
web3==6.15.1
httpx>=0.27.0

# Discord Bot
discord.py==2.7.1

# Telegram Bot
python-telegram-bot==21.0.1

# Twitter / X
tweepy[async]>=4.14.0

# Utilities
python-dotenv==1.0.1
apscheduler==3.10.4
